// src/app/shared/enum/status-order-enum.ts
export enum StatusOrderEnum {
  Created = 'Creada',
  Assigned = 'Asignada',
  AcceptedByTrade = 'Aceptada Por Comercio',
  RejectedByCommerce = 'Rechazada Por Comercio',
  Dispatched = 'Despachado',
  Delivered = 'Entregado',
  CancelByClient = 'Cancelada por cliente',
  ReceivedByClient = 'Recibida OK por cliente',
  DeliveryWithNews = 'Entrega con novedades cliente'
}